+++ lib/css.mli
@@ -4131,6 +4131,7 @@ type grid_line = Properties.grid_line =
   | Span_name of string  (** span <custom-ident> *)
   | Span_num_name of int * string  (** span <integer> <custom-ident> *)
   | Calc of grid_line calc  (** calc(12 * -1), etc. *)
+  | Calc_name of grid_line calc * string  (** calc(2) <custom-ident> *)
   | Var of grid_line var
 
 type grid_line_pair = Properties.grid_line_pair =